+++ src/main/java/roomescape/controller/ReservationController.java
@@ -3,8 +3,8 @@ package roomescape.controller;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import roomescape.dto.ReservationRequest;
-import roomescape.dto.ReservationResponse;
+import roomescape.controller.request.ReservationRequest;
+import roomescape.controller.response.ReservationResponse;
 import roomescape.service.ReservationService;
 
 @RestController
@@ -17,26 +17,19 @@ public class ReservationController {
     }
 
     @PostMapping
-    public ResponseEntity<Void> createReservation(
-            @RequestBody ReservationRequest reservationRequest
-    ) {
+    public ResponseEntity<Void> createReservation(@RequestBody ReservationRequest reservationRequest) {
         reservationService.createReservation(reservationRequest);
         return new ResponseEntity<>(HttpStatus.CREATED);
     }
 
     @GetMapping("/{reservationId}")
-    public ResponseEntity<ReservationResponse> getReservation(
-            @PathVariable Long reservationId
-    ) {
-        return ResponseEntity.ok(
-                reservationService.getReservation(reservationId)
-        );
+    public ResponseEntity<ReservationResponse> getReservation(@PathVariable Long reservationId) {
+        ReservationResponse reservationResponse = reservationService.getReservation(reservationId);
+        return ResponseEntity.ok(reservationResponse);
     }
 
     @DeleteMapping("/{reservationId}")
-    public ResponseEntity<Void> deleteReservation(
-            @PathVariable Long reservationId
-    ) {
+    public ResponseEntity<Void> deleteReservation(@PathVariable Long reservationId) {
         reservationService.deleteReservation(reservationId);
         return ResponseEntity.noContent().build();
     }
