+++ src/main/java/roomescape/controller/request/ReservationRequest.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.controller.request;
 
 import roomescape.domain.Reservation;
 import roomescape.domain.Theme;
